+++ CHANGELOG.md
@@ -53,6 +53,8 @@ These changes are available on the `master` branch, but have not yet been releas
   ([#2417](https://github.com/Pycord-Development/pycord/pull/2417))
 - `Option` may be used instead of `BridgeOption` until 2.7.
   ([#2417](https://github.com/Pycord-Development/pycord/pull/2417))
+- `Guild.query_members` now accepts `limit=None` to retrieve all members.
+  ([#2419](https://github.com/Pycord-Development/pycord/pull/2419))
 
 ## [2.5.0] - 2024-03-02